✅test: 자동 구매 확정 메서드 이동에 따른 테스트 코드 수정
